requirements.txt
asgiref==3.6.0
certifi==2023.7.22
charset-normalizer==3.2.0
dj-static==0.0.6
Django==4.2
djangorestframework==3.14.0
idna==3.4
mysqlclient==2.2.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.3
static3==0.7.0
tzdata==2023.3
urllib3==2.0.4
